requirements.txt
# Pour l'ingestion, les transformations et l'API
boto3
datasets
flask
pymysql
pymongo
python-dotenv
nltk

# Pour Airflow (vous pouvez aussi gérer la version via l'image Docker)
apache-airflow

# Pour le front-end Streamlit
streamlit
requests
protobuf

